Feat : 사용자 이미지 Div 구현 및 특정 Div 컴포넌트화
