--- src/components/Nav.tsx
@@ -1,5 +1,0 @@
-const Nav = () => {
-    return <div>NavPage</div>;
-};
-
-export default Nav;
