+++ src/components/nav/index.ts
@@ -1,0 +1,1 @@
+export { default as IconDiv } from './IconDiv';
