+++ src/components/project/Project.tsx
@@ -1,0 +1,5 @@
+const Project = () => {
+    return <div>Project</div>;
+};
+
+export default Project;
