+++ src/pages/index.tsx
@@ -2,6 +2,6 @@
 import 정리용 페이지
 import {HomePage, NavPage, PageNotFound} from './components'    
 */
-export { HomePage } from './Home';
-export { NavPage } from './Nav';
+export { HomePage } from './home';
+export { NavPage } from './nav';
 export { PageNotFound } from './PageNotFound';
